Consolidate UTF-8/Latin-1 helpers into encoding.c
Replace the static db_utf8_to_latin1_inplace clone in db.c and the
static latin1_to_utf8 in messages.c with a shared encoding module.
Tests now link the real implementation instead of relying on
duplicated bodies to dodge lang.c's other dependencies.
