+++ src/Makefile.am
@@ -23,6 +23,7 @@ XASTIR_SRC = \
     db_gui.c db_gui.h \
     dbfawk.c dbfawk.h \
     dlm.c dlm.h \
+    encoding.c encoding.h \
     dr_utils.c dr_utils.h \
     draw_symbols.c draw_symbols.h \
     fcc_data.c fcc_data.h \
