+++ tests/Makefile.am
@@ -24,7 +24,7 @@ test_interface_helpers_SOURCES = test_interface_helpers.c
 test_interface_helpers_CPPFLAGS = -I$(top_srcdir)/src
 
 
-test_db_SOURCES = test_db.c test_db_stubs.c $(top_srcdir)/src/db.c
+test_db_SOURCES = test_db.c test_db_stubs.c $(top_srcdir)/src/db.c $(top_srcdir)/src/encoding.c
 test_db_CPPFLAGS = $(CPPFLAGS) -I$(top_srcdir) -I$(top_srcdir)/src -I$(top_builddir)
 #test_db_LDADD = -L$(top_builddir)/src/rtree -lrtree
 
@@ -40,7 +40,7 @@ test_output_my_aprs_data_LDADD = -lpthread
 test_util_SOURCES = test_util.c test_util_stubs.c $(top_srcdir)/src/util.c
 test_util_CPPFLAGS = $(CPPFLAGS) -I$(top_srcdir) -I$(top_srcdir)/src -I$(top_builddir)
 
-test_objects_SOURCES = test_objects.c test_objects_stubs.c $(top_srcdir)/src/objects.c $(top_srcdir)/src/util.c $(top_srcdir)/src/object_utils.c $(top_srcdir)/src/db.c
+test_objects_SOURCES = test_objects.c test_objects_stubs.c $(top_srcdir)/src/objects.c $(top_srcdir)/src/util.c $(top_srcdir)/src/object_utils.c $(top_srcdir)/src/db.c $(top_srcdir)/src/encoding.c
 test_objects_CPPFLAGS = $(CPPFLAGS) -I$(top_srcdir) -I$(top_srcdir)/src -I$(top_builddir)
 
 # Nominatim tests (conditional on HAVE_NOMINATIM)
